application-code/kafka-fake-data-consumer2/requirements.txt
faker
kafka-python
psycopg2-binary

